hw5_Human_Face_Generation/requirements.txt
# Web scraping
requests
beautifulsoup4

# Date & time handling
pytz

# Image I/O and processing
pillow
imagehash
face_recognition
opencv-python

# Deep learning
torch>=2.0.0
torchvision
diffusers
transformers
accelerate
tqdm

# Others
numpy
scipy
